Prevent classname conflict
Related to this issue: https://github.com/nystudio107/craft-seomatic/issues/209

Since the class `Json` is already defined in the `\nystudio107\seomatic\helpers` namespace, PHP doesn't know which version to pick. Older PHP versions will pick the classname that is defined in the namespace instead of being overwritten by the `use`.